vs spinnt rum

+++ owncloud universal/Model/LinkStatus.cs
@@ -8,10 +8,10 @@ namespace owncloud_universal.Model
 {
     public class LinkStatus
     {
-        public int Id { get; set; }
-        public int TargetItemId { get; set; }
-        public int SourceItemId { get; set; }
-        public int ChangeNumber { get; set; }
-        public int AssociationId { get; set; }
+        public long Id { get; set; }
+        public long TargetItemId { get; set; }
+        public long SourceItemId { get; set; }
+        public long ChangeNumber { get; set; }
+        public long AssociationId { get; set; }
     }
 }
